requirements.txt
tensorflow>=2.10.0
streamlit>=1.20.0
numpy
pillow
matplotlib
seaborn
scikit-learn
opencv-python
h5py
